Handle MODE_DEBUG in main loop.

+++ source/frontends/common2/speed.cpp
@@ -19,11 +19,11 @@ namespace common2
     myStartCycles = g_nCumulativeCycles;
   }
 
-  size_t Speed::getCyclesTillNext(const size_t microseconds) const
+  uint64_t Speed::getCyclesTillNext(const size_t microseconds) const
   {
     if (myFixedSpeed)
     {
-      const size_t cycles = static_cast<uint64_t>(microseconds * g_fCurrentCLK6502 * 1.0e-6);
+      const uint64_t cycles = static_cast<uint64_t>(microseconds * g_fCurrentCLK6502 * 1.0e-6);
       return cycles;
     }
     else
